Added Auth functions

+++ studybud/base/views.py
@@ -1,6 +1,15 @@
 from django.shortcuts import render,redirect
 from django.template.context_processors import request
-from .models import Room
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.admin import register
+from django.template import Context
+from .models import Room, Topic
 from .forms import RoomForm
 
 # rooms= [
@@ -11,9 +20,62 @@ from .forms import RoomForm
 #     {'id':5, 'name':'Databases'},
 # ]
 
+def loginPage(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username= request.POST.get('username').lower()
+        password= request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username= username)
+        except:
+            messages.error(request,'User does not exist.')
+        
+        user = authenticate(request, username=username, password=password)    
+        
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        
+        else:
+            messages.error(request,'User OR password does not exist')
+        
+        
+    context= {'page':page}
+    return render(request, 'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page='register'
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form= UserCreationForm(request.POST)
+        
+        if form.is_valid():
+            user= form.save(commit=False)    
+            user.username =user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration!')
+            
+    context={'form':form}
+    return render(request, 'base/login_register.html',context)
+
 def home(request):
-    rooms= Room.objects.all()
-    context= {'rooms':rooms}
+    q= request.GET.get('q') if request.GET.get('q')!=None else ''
+    rooms= Room.objects.filter(Q(topic__name__icontains=q) | Q(name__icontains=q) | Q(description__icontains=q))
+    topics= Topic.objects.all()
+    room_count= rooms.count()
+    context= {'rooms':rooms, 'topics':topics, 'room_count':room_count}
+    
     # return render(request, 'home.html',{'rooms':rooms})
     return render(request, 'base/home.html',context)
 
@@ -22,6 +84,7 @@ def room(request,pk):
     context=   {'room':room} 
     return render(request, 'base/room.html',context) 
 
+@login_required(login_url='login')
 def createRoom(request):
     form= RoomForm()
     if request.method=='POST':
@@ -34,9 +97,13 @@ def createRoom(request):
     context= {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room= Room.objects.get(id=pk)
     form= RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allwed here!!')
     
     if request.method=='POST':
         form = RoomForm(request.POST, instance=room)
@@ -47,10 +114,17 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html',context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room= Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allwed here!!')
+    
     if(request.method=='POST'):
         room.delete()
         return redirect('home')
     
     return render(request, 'base/delete.html',{'obj':room})
+
